hw: arm: acpi: upgrade SPCR table to revision 4
On the ARM virt machine, there is currently no way to programmatically
discover the frequency of the UART reference clock solely through the use of
UEFI/ACPI (without the DTB). The SPCR table can include this information
as of revision 4, but is currently at revision 2.

Bump the revision to 4, add the clock frequency of 24 MHz and
complete the other new fields.

Signed-off-by: Vadim Chichikalyuk <chichikalyuk@gmail.com>